Created load-scripts helper
